test(extract): split tests per source file and raise coverage
